Update main.py
cleaned up and commented the code. Part of the code is commented out because it was used to format the .csv file and does not need to be run again. It does however show what changes were made to the file. 
The other part of the code will most likely be deleted because it is not needed and I am using the other .py file to create the DB. I need to go over it again.